requirements.txt
asgiref==3.9.1
django==5.2.4
django-cors-headers==4.7.0
django-guardian==3.0.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pyjwt==2.10.1
python-decouple==3.8
pytz==2025.2
pyyaml==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
